+++ packages/firebase-server/src/lib/env/env.ts
@@ -1,0 +1,6 @@
+/**
+ * Checks process.env.NODE_ENV for "test"
+ */
+export function isTestNodeEnv() {
+  return process.env.NODE_ENV === 'test';
+}