+++ packages/firebase-server/src/lib/index.ts
@@ -1,4 +1,6 @@
 export * from './auth';
+export * from './auth';
+export * from './env';
 export * from './firebase';
 export * from './firestore';
 export * from './function';