+++ packages/util/src/lib/array/array.string.spec.ts
@@ -1,4 +1,4 @@
-import { containsAllStringsAnyCase, containsAnyStringAnyCase, containsStringAnyCase, findUniqueCaseInsensitiveStrings, findUniqueTransform, TransformSingleStringFunction } from './array.string';
+import { containsAllStringsAnyCase, containsAnyStringAnyCase, containsStringAnyCase, findUniqueCaseInsensitiveStrings, findUniqueTransform, searchStringFilterFunction, TransformSingleStringFunction, caseInsensitiveFilterByIndexOfDecisionFactory } from './array.string';
 
 describe('findUniqueCaseInsensitiveStrings', () => {
   it('should return only the strings that are unique from the array.', () => {
@@ -134,6 +134,25 @@ describe('findUniqueTransform', () => {
       const result = transform(['test', 'TEST', 'tesT', 'TesT']);
       expect(result.length).toBe(1);
       expect(result[0]).toBe('TEST');
+    });
+  });
+});
+
+describe('searchStringFilterFunction', () => {
+  describe('function', () => {
+    const filterFunction = searchStringFilterFunction<string>({
+      readStrings: (x: string) => [x],
+      decisionFactory: caseInsensitiveFilterByIndexOfDecisionFactory
+    });
+
+    it('should return the values that match the filter', () => {
+      const filterText = 'a';
+      const values = ['a', 'AA', 'b'];
+
+      const result = filterFunction(filterText, values);
+      expect(result).toContain('a');
+      expect(result).toContain('AA');
+      expect(result).not.toContain('b');
     });
   });
 });